% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{add_path}
\alias{add_path}
\title{Add a path}
\usage{
add_path(con, user_id, original, hash)
}
\arguments{
\item{con}{Database connection.}

\item{user_id}{Id of the user adding the path.}

\item{original}{Original URL, URL being shortened.}

\item{hash}{Path of the shortened URL.}
}
\description{
Adds a record path to the database.
}
\keyword{internal}
