% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{allpairs_plot}
\alias{allpairs_plot}
\title{All Pairs Plots}
\usage{
allpairs_plot(x, y, type = "pca", rank, k = 0)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{type}{type of reduced-rank regression model to fit. \code{type = "identity"}, the default, uses \eqn{\mathbf{\Gamma} = \mathbf{I}} to fit a reduced-rank regression. \code{type = "pca"} fits a principal component analysis model as a special case of reduced-rank regression. \code{type = "cva"} fits a canonical variate analysis model as a special case of reduced-rank regression. \code{type = "lda"} fits a linear discriminant analysis model as a special case of reduced-rank regression.}

\item{rank}{rank of coefficient matrix.}

\item{k}{small constant added to diagonal of covariance matrices to make inversion easier.}
}
\value{
scatterplot matrix.
}
\description{
All Pairs Plots
}
\examples{
data(pendigits)
digits_features <- pendigits[, -35:-36]
digits_class <- pendigits[,35]
allpairs_plot(digits_features, digits_class, type = "pca", rank = 3)

}

