% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{annotation_custom_text}
\alias{annotation_custom_text}
\title{A function that extends ggplot2::annotation_custom by automatically
picking a textGrob, and allowing the user to pass the graphical parameters
of the grob into the function call itself.}
\usage{
annotation_custom_text(label, xmin = -Inf, xmax = Inf, ymin = -Inf,
  ymax = Inf, ...)
}
\arguments{
\item{label}{a string.}

\item{xmin}{x location (in data coordinates) giving horizontal
location of raster}

\item{xmax}{x location (in data coordinates) giving horizontal
location of raster}

\item{ymin}{y location (in data coordinates) giving vertical
location of raster}

\item{ymax}{y location (in data coordinates) giving vertical
location of raster}

\item{...}{optional, passed to grid::gpar.}
}
\description{
A function that extends ggplot2::annotation_custom by automatically
picking a textGrob, and allowing the user to pass the graphical parameters
of the grob into the function call itself.
}
