% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{authenticate_user}
\alias{authenticate_user}
\title{Authenticate a user}
\usage{
authenticate_user(con, email, password)
}
\arguments{
\item{con}{Database connection.}

\item{email, password}{User credentials.}
}
\description{
Authenticates a user; compares password hashes.
}
\keyword{internal}
