% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{binary_matrix}
\alias{binary_matrix}
\title{Binary Indicator Matrix}
\usage{
binary_matrix(class)
}
\arguments{
\item{class}{vector or data frame of type `character`}
}
\value{
matrix of binary indicator matrix where each column represents
}
\description{
Binary Indicator Matrix
}

