% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{canonical_corr}
\alias{canonical_corr}
\title{Canonical Correlation Matrix}
\usage{
canonical_corr(x, y, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{canonical_corr}
}

