% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{canonical_cov}
\alias{canonical_cov}
\title{Canonical Covariance Matrix}
\usage{
canonical_cov(x, y, rank = "full", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{canonical_cov}
}
\examples{
data(COMBO17)


}

