% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COMBO17}
\alias{COMBO17}
\title{MMST COMBO17 DATA}
\format{A data frame with 3462 observations on 65 numeric variables.}
\usage{
COMBO17
}
\description{
COMBO-17 galaxy photometric catalogue, 216, 219, 235
}
\references{
A. Izenman (2008). \emph{Modern Multivariate Statistical Techniques}. Springer.

Wolf, C. Meisenheimer, M., Kleinheinrich, M., Borch, A., Dye, S., Gray, M., Wisotski, L., Bell, E.F., Rix, H., W. Cimatti, A., Hasinger, G., and Szokoly, G. (2004). \emph{A catalogue of the Chandra Deep Field South with multi-colour classification and photometric redshifts from COMBO-17, Astronomy & Astrophysics}. \url{https://arxiv.org/pdf/astro-ph/0403666.pdf}
}
\keyword{datasets}

