% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_user}
\alias{create_user}
\title{Create a user}
\usage{
create_user(con, email, password)
}
\arguments{
\item{con}{Database connection.}

\item{email}{User's email.}

\item{password}{Raw password (this function encrypts it).}
}
\description{
Creates a new user record.
}
