% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cv_pairwise_plot}
\alias{cv_pairwise_plot}
\title{Pairwise Canonical Variates Plot}
\usage{
cv_pairwise_plot(x, y, cv_pair = 1, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{cv_pair}{integer. canonical variate pair to be plotted}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
Pairwise Canonical Variates Plot
}
\examples{
library(dplyr)
data(COMBO17)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cv_pairwise_plot(galaxy_x, galaxy_y)

}

