% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.R
\name{cv_scores}
\alias{cv_scores}
\title{Canonical Variate Scores}
\usage{
cv_scores(cva_object, x, y)
}
\arguments{
\item{cva_object}{A reduced-rank CVA object obtained by \code{cva}}

\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}
}
\description{
\code{cv_scores} calculates the canonical variate scores for \eqn{\mathbf{X}} and \eqn{\mathbf{Y}}
}

