% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.R
\name{cva}
\alias{cva}
\title{Fit Reduced-Rank CVA Model}
\usage{
cva(x, y, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\value{
list containing: matrix of means; matrix of canonical variate coefficients of predictor variables; matrix of canonical variate coefficients of response variables; vector of canonical correlations.
}
\description{
\code{cva} fits a reduced-rank canonical variate/correlation model. This is a special case of reduced-rank regression
with the weight matrix set to the covariance matrix of \eqn{Y}, i.e., \eqn{\mathbf{\Gamma} = \mathbf{\Sigma}_{YY}}.
Canonical variate analysis creates a set of new predictor variables that are linear combinations of the orignal predictors,
and a set of new resonse variables that are linear combinations of the original responses, such that each pair of new predictor and new
response maximizes correlation between the two and all pairs of canonical variates are independent of each other.
}
\examples{
library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cva(galaxy_x, galaxy_y, rank = 2)

}
\references{
Izenman, A. J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}
\seealso{
\code{\link{rrr}}
}

