% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{cva_error}
\alias{cva_error}
\title{Error of Reduced-Rank Canonical Variate Analysis}
\usage{
cva_error(x, y, x_new, y_new, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{x_new}{data frame or matrix of input variables used to predict the response matrix.}

\item{y_new}{data frame or matrix of observed response variables.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
Error of Reduced-Rank Canonical Variate Analysis
}

