% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cva_pairwise_plot}
\alias{cva_pairwise_plot}
\title{Pairwise Canonical Variates Plot}
\usage{
cva_pairwise_plot(x, y, cv_pair = 1, type = "cov", k = 0,
  interactive = TRUE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{cv_pair}{integer. canonical variate pair to be plotted}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{interactive}{if TRUE, creates an interactive plotly graphic.}
}
\value{
ggplot object if \code{interactive = FALSE}, the default, or an interactive plotly plot if \code{interactive = TRUE}.
}
\description{
For a given canonical-variate pair, \code{cva_pairwise_plot} produces a scatter plot with the canonical variate scores of the predictor
variables along the \eqn{X}-axis against the canonical-variate scores of the response variables along the \code{Y}-axis.
}
\examples{
library(dplyr)
data(COMBO17)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cva_pairwise_plot(galaxy_x, galaxy_y)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

