% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cva_rank_trace}
\alias{cva_rank_trace}
\title{Rank Trace Plot for Reduced-Rank CVA}
\usage{
cva_rank_trace(x, y, type = "cov", k = 0, plot = TRUE,
  interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{plot}{if FALSE, returns data frame of rank trace coordinates.}

\item{interactive}{if TRUE, creates an interactive plotly graphic.}
}
\value{
ggplot object if plot is \code{TRUE}, a data frame of rank trace coordinates if plot = \code{FALSE},
or an interactive plotly object if interactive = \code{TRUE}.
}
\description{
\code{cva_rank_trace} is a plot used to determine the effective dimensionality, i.e., \eqn{t = \mathrm{rank}\left(\mathbf{C}\right)},
of the reduced-rank regression equation, or the number of canonical variates to be used in the model.
}
\examples{
library(dplyr)
data(COMBO17)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cva_rank_trace(galaxy_x, galaxy_y)
cva_rank_trace(galaxy_x, galaxy_y, plot = FALSE)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}
\seealso{
\code{\link{rank_trace}}
}

