% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cva_rank_trace_plot}
\alias{cva_rank_trace_plot}
\title{Rank Trace Plot for Reduced-Rank CVA}
\usage{
cva_rank_trace_plot(x, y, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{cva_rank_trace_plot}
}

