% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cva_residual_3D_plot}
\alias{cva_residual_3D_plot}
\title{3D Plot of Residuals for Reduced-Rank CVA}
\usage{
cva_residual_3D_plot(x, y, cva_x = 1, cva_y = 2, cva_z = 3,
  rank = "full", type = "cov", k = 0, point_size = 3)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{cva_x}{integer number of the canonical variate used for the x-axis.}

\item{cva_y}{integer number of the canonical variate used for the y-axis.}

\item{cva_z}{integer number of the canonical variate used for the z-axis.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{point_size}{size of points in scatter.}
}
\value{
plotly object.
}
\description{
\code{cva_residual_3D_plot} creates an interactive, html plotly plot that can be manipulated by the viewer.
}

