% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{cva_residuals}
\alias{cva_residuals}
\title{Residuals of Reduced-Rank Canonical Variate Analysis}
\usage{
cva_residuals(x, y, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{cva_residuals} returns the multivariate residuals of the reduced-rank CVA regression
}
\examples{
library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_x <- select(galaxy, -Rmag:-chi2red)
galaxy_y <- select(galaxy, Rmag:chi2red)
cva_residuals(galaxy_x, galaxy_y, rank = 2, k = 0.001)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

