% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{delete.this}
\alias{delete.this}
\title{A function to automatically remove temporary objects from your .GlobalEnv.}
\usage{
delete.this(pattern = "nephew")
}
\arguments{
\item{pattern}{a regular expression passed to rm}
}
\description{
A function to automatically remove temporary objects from your .GlobalEnv.
}
