% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrR.R
\name{delete_this}
\alias{delete_this}
\title{A function to automatically remove temporary objects from your .GlobalEnv.}
\usage{
delete_this(pattern = "nephew")
}
\arguments{
\item{pattern}{a regular expression passed to rm}
}
\description{
A function to automatically remove temporary objects from your .GlobalEnv.
}
