% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{diffr}
\alias{diffr}
\title{A version of setdiff with a "reverse order" option and a "cbind" version
that combines the "regular" and the "reverse" order into a data.frame.}
\usage{
diffr(x, y, order = "xy")
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{order}{One of "xy", "yx", or "df".}
}
\value{
a vector of items present in X (or Y) but not in Y (or X)
}
\description{
A version of setdiff with a "reverse order" option and a "cbind" version
that combines the "regular" and the "reverse" order into a data.frame.
}
