% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expand_feature_space}
\alias{expand_feature_space}
\title{Expand Feature Space}
\usage{
expand_feature_space(feature_space)
}
\arguments{
\item{feature_space}{a matrix or data frame of input variables}
}
\value{
data frame of all input variables, their squares and all cross products
}
\description{
\code{expand_feature_space}
}
\examples{
data(iris)
iris_features <- iris[,1:4]
expand_feature_space(iris)

}

