% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table.r
\name{get.table.symmetric}
\alias{get.table.symmetric}
\title{Calculate table of pair (n1, n2) with a given minimal power for symmetric design}
\usage{
get.table.symmetric(powertables, k = 3)
}
\arguments{
\item{powertables}{A list produced by the \code{calculate.powertables()}}

\item{k}{number of active treatment groups to be tested}
}
\value{
A data frame combinations of (n1, n2)
}
\description{
Calculate table of pair (n1, n2) with a given minimal power for symmetric design
}

