% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_hash_data}
\alias{get_hash_data}
\title{Retrieve data}
\usage{
get_hash_data(con, hash)
}
\arguments{
\item{con}{Database connection.}

\item{hash}{Path to retrieve data for.}
}
\description{
Retrieve the data of a given URL.
}
\keyword{internal}
