% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_path}
\alias{get_path}
\title{Retrieve a URL record}
\usage{
get_path(con, path)
}
\arguments{
\item{con}{Database connection.}

\item{path}{Path to retrieve record.}
}
\description{
Retrieve a URL record given its path.
}
\keyword{internal}
