% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_user}
\alias{get_user}
\title{Get user by email}
\usage{
get_user(con, email)
}
\arguments{
\item{con}{Database connection.}

\item{email}{Email address.}
}
\description{
Retrieve a user record given its email address.
}
\keyword{internal}
