% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_user_by_id}
\alias{get_user_by_id}
\title{Retrieve a User}
\usage{
get_user_by_id(con, id)
}
\arguments{
\item{con}{Database connection.}

\item{id}{Id of user, an integer.}
}
\description{
Retrieve a user record given it's unique id.
}
\keyword{internal}
