% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrR.R
\name{global_size}
\alias{global_size}
\title{A function that prints a named vector with the object_sizes of the objects
in the .GlobalEnv.}
\usage{
global_size()
}
\description{
A function that prints a named vector with the object_sizes of the objects
in the .GlobalEnv.
}
