% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{handler}
\alias{handler}
\title{Inherit params}
\arguments{
\item{res}{An object of class `Response` as obtained by
[ambiorix::Response].}

\item{req}{An object of class `Request` as obtained by
[ambiorix::Request].}
}
\description{
Inherit params
}
