% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{increment}
\alias{increment}
\alias{increment_data}
\alias{update_data}
\alias{create_data}
\title{Increment count}
\usage{
increment_data(con, path)

update_data(con, date, id, count)

create_data(con, date, id)
}
\arguments{
\item{con}{Database connection.}

\item{path}{Path to increment}
}
\description{
Increment the count for a given path.
}
\details{
This is used to increment the `data` table at
every redirect. 
We only store the `url_id`, the `date`, and the `count`.
}
\keyword{internal}
