% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{is_authenticated}
\alias{is_authenticated}
\title{Auth Check}
\usage{
is_authenticated(con, cookie)
}
\arguments{
\item{cookie}{Cookie as returned by `req$cookie$rrr`.}
}
\description{
Checks whether a user is actually authenticated.
}
\keyword{Database}
\keyword{con}
\keyword{connection.}
\keyword{internal}
