% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{is_email}
\alias{is_email}
\title{Is Email}
\usage{
is_email(email)
}
\arguments{
\item{email}{Email address to check.}
}
\value{
boolean/logical.
}
\description{
Basic regex check to ensure the email address is valid.
}
\keyword{internal}
