% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{is_valid_path}
\alias{is_valid_path}
\title{Valid Path}
\usage{
is_valid_path(path)
}
\arguments{
\item{path}{Path to check}
}
\value{
bool/logical.
}
\description{
Ensures we have a valid path.
We only accept alphanumeric.
}
\keyword{internal}
