% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_plots.R
\name{ld_pairwise_plot}
\alias{ld_pairwise_plot}
\title{LD Pairwise Plot}
\usage{
ld_pairwise_plot(x, class, ld_x = 1, ld_y = 2, rank = "full",
  type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{class}{vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.}

\item{ld_x}{integer. Linear discriminant function plotted along the x-axis.}

\item{ld_y}{integer. Linear discriminant function plotted along the y-axis.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{ld_pairwise_plot}
}

