% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Fit Reduced-Rank LDA Model}
\usage{
lda(x, class, rank = "full", type = "cov", k = 0, quadratic = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{class}{vector or one-column data frame of type factor or character that are the class labels of the observations.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{quadratic}{logical. If \code{TRUE} fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.}
}
\value{
list containing: a named vector of prior probabilities; a data frame of canonical-variate coefficients of the predictor variables;
a data frame of canonical-variate coefficients for the reponse variables.
}
\description{
\code{lda} fits a reduced-rank linear discriminant analysis model by fitting a reduced-rank canonical variate
model to the predictor variables and a response-indicator matrix.
}
\examples{
data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
lda(iris_x, iris_y)

}
\references{
Izenman, A. J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

