% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_plots.R
\name{lda_pairwise_plot}
\alias{lda_pairwise_plot}
\title{LDA Pairwise Plot}
\usage{
lda_pairwise_plot(x, class, lda_x = 1, lda_y = 2, rank = "full",
  type = "cov", k = 0, quadratic = FALSE, interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{class}{vector or one-column data frame of type factor or character that are the class labels of the observations.}

\item{lda_x}{integer. Linear discriminant function plotted along the x-axis.}

\item{lda_y}{integer. Linear discriminant function plotted along the y-axis.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{quadratic}{logical. If \code{TRUE} fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.}

\item{interactive}{logical. If \code{TRUE} prints an interactive Plotly graphic.}
}
\value{
ggplot object if \code{interactive = FALSE}, the default, or an interactive plotly plot if \code{interactive = TRUE}.
}
\description{
\code{lda_pairwise_plot} creates a plot with the linear discriminant scores of one linear discriminant
function along the \eqn{X}-axis against the linear discriminant scores of another linear discriminant function
alonge the \code{Y}-axis.
}
\examples{
data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
lda_pairwise_plot(iris_x, iris_y)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

