% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda_scores}
\alias{lda_scores}
\title{Linear Discriminant Scores}
\usage{
lda_scores(x, class, rank = "full", type = "cov", k = 0,
  quadratic = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{class}{vector or one-column data frame of type factor or character that are the class labels of the observations.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{quadratic}{logical. If \code{TRUE} fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.}
}
\value{
list containing: data frame of linear discriminant function scores; data frame of class means.
}
\description{
\code{lda_scores} calculates canonical-variate scores. In the case of linear discriminant analysis, the canonical-variate
scores of the predictor variables are the linear discriminant functions evaluated at the values of predictor variables, while the scores for the
response variables correspond to class means.
}
\examples{
data(iris)
lda_scores(iris[,1:4], iris[,5])

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

