% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout_get}
\alias{logout_get}
\title{GET Logout}
\usage{
logout_get(req, res)
}
\arguments{
\item{req}{A `Request`.}

\item{res}{A `Response`.}
}
\description{
Logout, clears cookie and redirects to homepage.
}
\keyword{internal}
