% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{mid_tmpl_register}
\alias{mid_tmpl_register}
\title{Middleware}
\usage{
mid_tmpl_register(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
We could use multiple middleware functions...
}
\keyword{internal}
