% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{object_size}
\alias{object_size}
\title{A "prettified" object.size that returns sizes formatted in Mb. The point
here is exclusively, so far, to avoid having to type
'format(object.size(x), 'Mb') so much.}
\usage{
object_size(object)
}
\arguments{
\item{object}{an R object.}
}
\description{
A "prettified" object.size that returns sizes formatted in Mb. The point
here is exclusively, so far, to avoid having to type
'format(object.size(x), 'Mb') so much.
}
