% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{organize}
\alias{organize}
\title{Organize Data for RRR}
\usage{
organize(vars, scale = FALSE)
}
\arguments{
\item{vars}{data frame or matrix}

\item{scale}{logical. If \code{TRUE}, matrix is scaled by standard deviations.}
}
\description{
\code{organize} mean-center and transpose a matrix
}

