% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{path_exists}
\alias{path_exists}
\title{Path exists}
\usage{
path_exists(con, path)
}
\arguments{
\item{con}{Database connection.}

\item{path}{Path to check.}
}
\value{
boolean, logical.
}
\description{
Checks whether a path exists.
}
\keyword{internal}
