% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pc_allpairs_plot}
\alias{pc_allpairs_plot}
\title{Scatterplot Matrix of Principle Components}
\usage{
pc_allpairs_plot(x, rank, type = "cov", k = 0, class_labels = NULL)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}

\item{class_labels}{data frame or vector of class labels}
}
\description{
Scatterplot Matrix of Principle Components
}

