% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pc_plot_3D}
\alias{pc_plot_3D}
\title{3D Principal Component Plot}
\usage{
pc_plot_3D(x, pc_x = 1, pc_y = 2, pc_z = 3, class_labels = NULL,
  rank = "full", type = "cov")
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{pc_x}{integer number of the principal component used for the x-axis}

\item{pc_y}{integer number of the principal component used for the y-axis}

\item{pc_z}{integer number of the principal component used for the z-axis}

\item{class_labels}{data frame or vector of class labels}

\item{rank}{rank of coefficient matrix}

\item{type}{type of covariance matrix}
}
\description{
\code{pc_plot_3D}
}

