% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_3D_plot}
\alias{pca_3D_plot}
\title{3D Principal Component Plot}
\usage{
pca_3D_plot(x, pca_x = 1, pca_y = 2, pca_z = 3, class_labels = NULL,
  rank = "full", type = "cov", point_size = 3)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{pca_x}{integer number of the principal component used for the x-axis}

\item{pca_y}{integer number of the principal component used for the y-axis}

\item{pca_z}{integer number of the principal component used for the z-axis}

\item{class_labels}{data frame or vector of class labels}

\item{rank}{rank of coefficient matrix}

\item{type}{type of covariance matrix}

\item{point_size}{size of points in scatter}
}
\description{
code{pca_3D_plot}
}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

