% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_diagnostics.R
\name{pca_gof}
\alias{pca_gof}
\title{PCA Goodness of Fit}
\usage{
pca_gof(x)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}
}
\description{
\code{pca_gof} is a measure of the goodness of fit of the \eqn{i}th principal component.
}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

