% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_pairwise_plot}
\alias{pca_pairwise_plot}
\title{PC Pairwise Plot}
\usage{
pca_pairwise_plot(x, pc_x = 1, pc_y = 2, class_labels = NULL,
  rank = "full", type = "cov", interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{pc_x}{principal component for the x-axis}

\item{pc_y}{principal component for the y-axis}

\item{class_labels}{data frame or vector of class labels}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{interactive}{logical. If \code{TRUE} prints an interactive Plotly graphic.}
}
\description{
\code{pca_pairwise_plot}
}
\examples{
data(pendigits)
digits_features <- pendigits[,1:34]
digits_class <- pendigits[,35]
pca_pairwise_plot(digits_features, pc_x = 1, pc_y = 3)

}

