% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_plot_3D}
\alias{pca_plot_3D}
\title{3D Principal Component Plot}
\usage{
pca_plot_3D(x, pca_x = 1, pca_y = 2, pca_z = 3, class_labels = NULL,
  rank = "full", type = "cov")
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{pca_x}{integer number of the principal component used for the x-axis}

\item{pca_y}{integer number of the principal component used for the y-axis}

\item{pca_z}{integer number of the principal component used for the z-axis}

\item{class_labels}{data frame or vector of class labels}

\item{rank}{rank of coefficient matrix}

\item{type}{type of covariance matrix}
}
\description{
code{pca_plot_3D}
}

