% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_rank_trace.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{Rank Trace for PCA}
\usage{
pca_rank_trace(x)
}
\arguments{
\item{x}{export pca_rank_trace}
}
\description{
\code{pca_rank_trace} is a wrapper function for \code{rank_trace} for Principle Components Analysis
}

