% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_rank_trace_plot}
\alias{pca_rank_trace_plot}
\title{PCA Rank Trace Plot}
\usage{
pca_rank_trace_plot(x, type = "cov", k = 0, interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{interactive}{logical. If \code{TRUE} prints an interactive Plotly graphic.}
}
\description{
Plot of rank trace to determine number of principle components to use in reduced-rank PCA.
}
\examples{
data(pendigits)
digits_features <- pendigits[,1:34]
pca_rank_trace_plot(digits_features)

}
\seealso{
\code{pca_rank_trace} \code{rank_trace_plot}
}

