% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_residuals}
\alias{pca_residuals}
\title{Reduced-Rank PCA Residuals}
\usage{
pca_residuals(x, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{pca_residuals}
}

