% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pendigits}
\alias{pendigits}
\title{MMST PENDIGITS DATA}
\format{a data frame with 10992 observations on 36 unnamed variables}
\source{
\url{http://archive.ics.uci.edu/ml/datasets.html}
}
\usage{
pendigits
}
\description{
pen-based handwritten digit recognition, 211, 234, 274, 348, 391, 631
}
\references{
A. Izenman (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}
\keyword{datasets}

