% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Profile Router}
\usage{
profile(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Router for convenience and easier organisation
on the profile.
}
\keyword{internal}
