% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_get}
\alias{profile_get}
\title{GET Profile}
\usage{
profile_get(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Returns the profile page, unless the user is not
authenticated in which case it redirects to the login
page.
}
\keyword{internal}
