% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_post}
\alias{profile_post}
\title{POST Profile}
\usage{
profile_post(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Handle form POSTed from /profile, namely the
URL shortening.
}
\keyword{internal}
