% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_plots.R
\name{rank_trace}
\alias{rank_trace}
\title{Rank Trace}
\usage{
rank_trace(x, y, gamma_matrix)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}
}
\description{
Rank trace to determine number of principle components to use in reduced-rank PCA.
}

