% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_get}
\alias{register_get}
\title{GET Register}
\usage{
register_get(req, res)
}
\arguments{
\item{req}{A `Request`.}

\item{res}{A `Response`.}
}
\description{
Returns register page, unless the user is 
authenticated in which case we redirect to
the profile.
}
\keyword{internal}
