\name{reset.globals}
\alias{reset.globals}
\title{Resets the global variables}
\usage{
  reset.globals()
}
\value{
  Nothing. This function is called for its side effects in
  the \code{.GlobalEnv}.
}
\description{
  This function is called upon initialization to set the
  proper global variables.
}

