% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{reverse_names}
\alias{reverse_names}
\title{A (largely untested) function that will take a named vector or a named list
and reverse elements/names: the names become the elements, and the elements
become the names. This is useful to me because I can never remember the
structure of named lists preferred by the custom.coef.map argument in texreg.}
\usage{
reverse_names(veclist, coerce = FALSE)
}
\arguments{
\item{veclist}{a named vector or list.}

\item{coerce}{coerce a list to a vector? silently ignored if veclist is a 
vector.}
}
\description{
A (largely untested) function that will take a named vector or a named list
and reverse elements/names: the names become the elements, and the elements
become the names. This is useful to me because I can never remember the
structure of named lists preferred by the custom.coef.map argument in texreg.
}
