% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge Regression}
\usage{
ridge(x, y, k = 0)
}
\arguments{
\item{x}{data frame of input variables}

\item{y}{vector of response variables}

\item{k}{small constant to augment the diagonal of the covariance matrix \eqn{\Sigma_{XX}}}
}
\description{
\code{ridge} fits a generalized ridge regression model by minimizing sum of squared error subject to an elliptical restriction on model parameters.
}

