% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr.R
\name{rprop_table}
\alias{rprop_table}
\title{A shortcut for prop_table that displays values as percentages and rounds
them. For example, rprop_table(var, 1) is equivalent to 
round(prop_table(var)*100, 1).}
\usage{
rprop_table(..., digits = 1)
}
\arguments{
\item{...}{arguments to be passed to methods.}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}
}
\description{
A shortcut for prop_table that displays values as percentages and rounds
them. For example, rprop_table(var, 1) is equivalent to 
round(prop_table(var)*100, 1).
}
