% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{rrlda}
\alias{rrlda}
\title{Reduced-Rank Linear Discriminant Analysis}
\usage{
rrlda(x, y, k = 0, ld_num = 2)
}
\arguments{
\item{x}{data frame of input variables}

\item{y}{data frame of categorical response variables}

\item{k}{parameter to ensure non-singular covariance matrices}

\item{ld_num}{number of linear discriminants to split the data}
}
\description{
\code{lda} produces a linear discriminant analysis as a reduced-rank regression.
}

