% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{rrpca}
\alias{rrpca}
\title{Fit Reduced-Rank PCA Model}
\usage{
rrpca(x, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} is standard multivariate regression.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{pca} fits reduced-rank principle component analysis model.
}
\references{
Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.
}

