% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{rrpca_error}
\alias{rrpca_error}
\title{Reduced-Rank PCA Error}
\usage{
rrpca_error(rrpca_object, x_new)
}
\arguments{
\item{rrpca_object}{`list` object obtained from `rrpca()`}

\item{x_new}{data frame or matrix of new observations to predict.}
}
\description{
\code{rrpca_error}
}

