% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{rrpca_predict}
\alias{rrpca_predict}
\title{Predict via Reduced-Rank Principle Components Analysis}
\usage{
rrpca_predict(rrpca_object, x_new)
}
\description{
\code{rrpca_predict}
}
\details{
inheritParams rrpca
}

