\name{RRR-package}
\alias{RRR-package}
\alias{RRR}
\docType{package}
\title{
\packageTitle{RRR}
}
\description{
\packageDescription{RRR}
}
\details{
Main functions for multivariate regression: \code{\link{rrr}} (reduced rank ridge regression), 
\code{\link{rr}} (reduced rank regression), \code{\link{ridge}} (ridge regression with single tuning parameter), \code{\link{pls}} (partial least squares regression), \code{\link{pcr}} (principal component regression). K-fold cross-validation for selecting tuning parameters: \code{\link{kcv}}.

Main functions for simulating multivariate regression models: \code{\link{simulate.rr.model}}, \code{\link{simulate.pls.model}}, \code{\link{simulate.pcr.model}}. 
}
\author{
\packageAuthor{RRR}

Maintainer: \packageMaintainer{RRR}
}
\references{
Ashin Mukherjee and Ji Zhu. \emph{Reduced rank ridge regression and its kernel extensions.} Statistical analysis and data mining, 4(6):612–622, 2011.\cr
Gregory C. Reinsel, Raja P. Velu, and Kun Chen. \emph{Multivariate Reduced-Rank Regression. Theory, Methods and Applications.} Springer, 2022. \cr
\url{https://en.wikipedia.org/wiki/Ridge_regression}\cr
\url{https://en.wikipedia.org/wiki/Partial_least_squares_regression} \cr
\url{https://en.wikipedia.org/wiki/Principal_component_regression}
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = RRR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{RRR}}
%\section{Documentation Index}{\packageIndices{RRR}}

