% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr}
\alias{rrr}
\title{Reduced-Rank Regression}
\usage{
rrr(x, y, gamma_matrix, rank)
}
\arguments{
\item{x}{data frame of input variables}

\item{y}{data frame of response variables}

\item{rank}{of the coefficient matrix to estimate. \code{rank = full} is standard multivariate regression.}

\item{gamma_mat}{weight matrix}
}
\description{
\code{rrr} fits a reduced-rank regression model.
}

