% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_diagnostics.R
\name{rrr_error}
\alias{rrr_error}
\title{Reduced-Rank Regression Error}
\usage{
rrr_error(rrr_object, x_new, y_new)
}
\arguments{
\item{rrr_object}{an object of type `list` that contains the means and coefficients of the reduced-rank regression.}

\item{x_new}{data frame or matrix of input variables used to predict the response matrix.}

\item{y_new}{data frame or matrix of observed response variables.}
}
\description{
\code{rrr_error} calculates the error from predicting response variables from a test set using coefficients calculated from training data.
}
\examples{
data(tobacco)
set.seed(123)
train_index <- sample(dim(tobacco)[1], 18)
tobacco_x <- tobacco[,4:9]
tobacco_y <- tobacco[,1:3]
tobacco_train_x <- tobacco_x[train_index, ]
tobacco_train_y <- tobacco_y[train_index, ]
tobacco_test_x <- tobacco_x[-train_index, ]
tobacco_test_y <- tobacco_y[-train_index, ]
gamma <- diag(1, dim(tobacco_y)[2])
tobacco_rrr <- rrr(tobacco_train_x, tobacco_train_y, gamma, rank = 1)
rrr_error(tobacco_rrr, tobacco_test_x, tobacco_test_y)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

