% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_plots.R
\name{rrr_residual_plot}
\alias{rrr_residual_plot}
\title{Plot Residuals of Reduced-Rank Regression}
\usage{
rrr_residual_plot(x, y, gamma_matrix, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\value{
ggplot object, scatterplot matrix.
}
\description{
\code{rrr_residual_plot} is a scatter plot matrix used for diagnostics of the reduced-rank regression model.
}
\examples{
data(tobacco)
tobacco_x <- tobacco[,4:9]
tobacco_y <- tobacco[,1:3]
gamma <- diag(1, dim(tobacco_y)[2])
rrr_residual_plot(tobacco_x, tobacco_y, gamma)
gamma2 <- solve(cov(tobacco_y))
rrr_residual_plot(tobacco_x, tobacco_y, gamma2)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

