% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_diagnostics.R
\name{rrr_residuals}
\alias{rrr_residuals}
\title{RRR Residuals}
\usage{
rrr_residuals(x, y, gamma_matrix, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{rrr_residuals}
}
\examples{
data(tobacco)
tobacco_x <- tobacco[,4:9]
tobacco_y <- tobacco[,1:3]
gamma <- diag(1, dim(tobacco_y)[2])
tobacco_rrr <- rrr(tobacco_x, tobacco_y, gamma, rank = 1)
rrr_residuals(tobacco_x, tobacco_y, gamma, rank = 1)

}

