% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_redirect}
\alias{shorten_redirect}
\title{Redirect shroten}
\usage{
shorten_redirect(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Redirects shortened links if found.
Otherwise we return 404.
}
\keyword{internal}
