% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{shortened_path}
\alias{shortened_path}
\title{Create Shortened URL}
\usage{
shortened_path(path)
}
\arguments{
\item{path}{Path to create URL with.}
}
\description{
Create shortened URL from path.
}
\keyword{internal}
