% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sqrt_matrix}
\alias{sqrt_matrix}
\title{Square Root Matrix}
\usage{
sqrt_matrix(matr)
}
\arguments{
\item{matr}{a positive semi-definite matrix}
}
\description{
\code{sqrt_matrix} calculates the square root of a positive semi-definite matrix through spectral decomposition.
}

