% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{template_data}
\alias{template_data}
\title{Template Data}
\usage{
template_data(con, req, ...)
}
\arguments{
\item{con}{Database connection.}

\item{req}{A `Request`.}

\item{...}{Additional data (key value pairs).}
}
\description{
All `template_*` methods added in `middleware`
use this same function.
Useful to add data that is required in every 
(or most) templates, e.g.: whether
the user is authenticated.
}
\keyword{internal}
