% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{template_path}
\alias{template_path}
\title{Template Path}
\usage{
template_path(...)
}
\arguments{
\item{...}{Path, character.}
}
\description{
Create a path to the `templates` in `inst/templates`.
}
\keyword{internal}
