% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco}
\alias{tobacco}
\title{MMST TOBACCO DATA}
\format{a data frame with 25 observations on the following 9 variables.
\itemize{
  \item{‘Y1.BurnRate’ a numeric vector}
  \item{‘Y2.PercentSugar’ a numeric vector}
  \item{‘Y3.PercentNicotine’ a numeric vector}
  \item{‘X1.PercentNitrogen’ a numeric vector}
  \item{‘X2.PercentChlorine’ a numeric vector}
  \item{‘X3.PercentPotassium’ a numeric vector}
  \item{‘X4.PercentPhosphorus’ a numeric vector}
  \item{‘X5.PercentCalcium’ a numeric vector}
  \item{‘X6.PercentMagnesium’ a numeric vector}
  }}
\usage{
tobacco
}
\description{
chemical composition of tobacco, 183, 187
}
\references{
A. Izenman (2008). \emph{Modern Multivariate Statistical Techniques}. Springer.

Anderson, R.L. and Bancroft, T.A. (1952). \emph{Statistical Theory in Research}. New York: Mcgraw-Hill.
}
\keyword{datasets}

