% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{user_exists}
\alias{user_exists}
\title{User exists}
\usage{
user_exists(con, email)
}
\arguments{
\item{con}{Database connection.}

\item{email}{Email of the user.}
}
\description{
Checks whether a user exists.
}
\keyword{internal}
