% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{user_update_password}
\alias{user_update_password}
\title{Update a user password}
\usage{
user_update_password(con, id, password_hash)
}
\arguments{
\item{con}{Database connection.}

\item{id}{Id of the user.}

\item{password_hash}{Hashed password as returned by [password_encrypt()].}
}
\description{
Updates a user's password.
}
\keyword{internal}
