% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{views}
\alias{views}
\alias{home_get}
\title{Home}
\usage{
home_get(req, res)
}
\arguments{
\item{req}{An object of class `Request` as obtained by
[ambiorix::Request].}

\item{res}{An object of class `Response` as obtained by
[ambiorix::Response].}
}
\description{
Render the homepage.
}
\keyword{internal}
