% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrR.R
\name{which_is_class}
\alias{which_is_class}
\title{Find all objects in .GlobalEnv of a given class.}
\usage{
which_is_class(class_name)
}
\arguments{
\item{class_name}{a string, or a vector or list of strings.}
}
\description{
A function to quickly list all objects in your global environment of a given
class. Useful to me for one main reason: you can pass it to rm, with
rm(list = which_is_class('something something'))
This is quite convenient if, for example, you have a database connection,
assign a bunch of tables to objects, and later want to remove all of those
because you're done pulling data, because you closed the connection,
or because you've quit the project in a fit of rage.
}
