% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrR.R
\name{wtable}
\alias{wtable}
\title{A cousin of table(), appropriate for piping.}
\usage{
wtable(.data, ...)
}
\arguments{
\item{.data}{a data frame, tibble, data table, etc. if the object works with
the with() function, it works here.}

\item{...}{variables/arguments to be passed to table().}
}
\description{
This function -- identical to with() except for a direct call to table()
allows you to pipe to the table function. In practice, this means that you
can write df %>% wtable(variable) instead of df %>% {table(.$variable)}, or
df %>% with(table(variable)), or even df %$% table(variable). You might find
this function quite dumb -- and I don't blame you! -- but it does make things
slightly faster for me.
}
