% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{cva_error}
\alias{cva_error}
\title{Error of Reduced-Rank Canonical Variate Analysis}
\usage{
cva_error(x, y, x_new, y_new, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{y}{data frame or matrix of response variables}

\item{x_new}{data frame or matrix of input variables used to predict the response matrix.}

\item{y_new}{data frame or matrix of observed response variables.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\value{
data frame of error.
}
\description{
\code{cva_error} calculates the errors of a canonical-variate regression model built on a training set and applied to a test set.
}
\examples{
set.seed(12345)
library(dplyr)
data(COMBO17)
galaxy <- as_data_frame(COMBO17)
galaxy <- select(galaxy, -starts_with("e."), -Nr, -UFS:-IFD)
galaxy <- na.omit(galaxy)
galaxy_train <- 
sample_size <- floor(0.66 * nrow(galaxy))
train_ind <- sample(seq_len(nrow(galaxy)), size = sample_size)
train <- galaxy[train_ind, ]
test <- galaxy[-train_ind, ]
train_x <- select(train, -Rmag:-chi2red)
train_y <- select(train, Rmag:chi2red)
test_x <- select(test, -Rmag:-chi2red)
test_y <- select(test, Rmag:chi2red)
cva_error(train_x, train_y, test_x, test_y, rank = 2, k = 0.001)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

