% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_diagnostics.R
\name{cva_error}
\alias{cva_error}
\title{Error of Reduced-Rank Canonical Variate Analysis}
\usage{
cva_error(cva_object, x_new, y_new)
}
\arguments{
\item{cva_object}{cva object}

\item{x_new}{new x}

\item{y_new}{new y}
}
\description{
Error of Reduced-Rank Canonical Variate Analysis
}

