% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva_plots.R
\name{cva_residual_plot}
\alias{cva_residual_plot}
\title{Residual Plots for Reduced-Rank CVA}
\usage{
cva_residual_plot(x, y, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{cv_residual_plot}
}

