% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.contours.r
\name{get.contours}
\alias{get.contours}
\title{Calculate contours of pair (n1, n2) with a given minimal power}
\usage{
get.contours(z, power = 0.8)
}
\arguments{
\item{z}{A data frame produced by \code{power.grid()}}

\item{power}{Power of the test}
}
\value{
A data frame combinations of (n1, n2)
}
\description{
Calculate contours of pair (n1, n2) with a given minimal power
}
