% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeRmd.R
\name{includeRmd}
\alias{includeRmd}
\title{Compiles Rmd and output html}
\usage{
includeRmd(file, fragment.only = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{file}{file name of the .Rmd file}

\item{fragment.only}{logical passed to the knitr::knit2html() function}

\item{quiet}{Default \code{quiet = TRUE} passed down to \code{knit2html()}}

\item{\dots}{Additonal arguments passed down to \code{knit2html()}}
}
\value{
html vector
}
\description{
Compiles Rmd and output html
}
