% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Reduced-Rank Linear Discriminant Analysis}
\usage{
lda(x, y, rank = "full")
}
\arguments{
\item{rank}{}
}
\description{
\code{lda} produces a linear discriminant analysis as a reduced-rank regression.
}

