% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Fit Reduced-Rank LDA Model}
\usage{
lda(x, class, rank = "full", type = "cov", k = 0, quadratic = FALSE)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{class}{vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{quadratic}{`logical`. If `TRUE` fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.}
}
\value{
`list` containing: a data frame of class
}
\description{
\code{lda} fits a reduced-rank linear discriminant analysis model.
}
\examples{
data(pendigits)
digits_features <- pendigits[,-35:-36]
digits_class <- pendigits[,35]
lda(digits_features, digits_class, rank = 3, quadratic = TRUE)

}
\references{
Izenman, A. J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

