% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_plots.R
\name{lda_pairwise_plot}
\alias{lda_pairwise_plot}
\title{LDA Pairwise Plot}
\usage{
lda_pairwise_plot(x, class, lda_x = 1, lda_y = 2, rank = "full",
  type = "cov", k = 0, quadratic = FALSE, interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{class}{vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.}

\item{lda_x}{integer. Linear discriminant function plotted along the x-axis.}

\item{lda_y}{integer. Linear discriminant function plotted along the y-axis.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{quadratic}{`logical`. If `TRUE` fits a reduced-rank linear discriminant model on an expanded feature space that includes all the input variables, their squares, and all of pairwise cross-products.}
}
\description{
\code{lda_pairwise_plot}
}
\examples{
data(iris)
iris_x <- iris[,1:4]
iris_y <- iris[5]
lda_pairwise_plot(iris_x, iris_y)

}

