% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda_scores}
\alias{lda_scores}
\title{Linear Discriminant Scores}
\usage{
lda_scores(x, class, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{class}{vector or one-column data frame of type `factor` or `character` that are the class labels of the observations.}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{lda_scores}
}
\examples{
data(iris)
lda_scores(iris[,1:4], iris[,5])

}

