% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{organize}
\alias{organize}
\title{Organize Matrix for Reduced-Rank Regression}
\usage{
organize(vars, type = "cov")
}
\arguments{
\item{vars}{data frame or matrix of variables to be organized.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}
}
\description{
\code{organize} transposes and mean-centers a data frame or matrix of input or output variables.
}

