% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{organize}
\alias{organize}
\title{Organize data for RRR}
\usage{
organize(vars, scale = FALSE)
}
\value{
centered matrix whose rows are variables and columns observations
}
\description{
Organize data in format suitable for reduced-rank regression
}

