% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pc_pairwise_plot}
\alias{pc_pairwise_plot}
\title{PC Pairwise Plot}
\usage{
pc_pairwise_plot(x, pc_1 = 1, pc_2 = 2, class_labels = NULL,
  rank = "full", type = "cov", interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{pc_1}{principal component for the x-axis}

\item{pc_2}{principal component for the y-axis}

\item{class_labels}{data frame or vector of class labels}

\item{rank}{rank of coefficient matrix}

\item{type}{type of covariance matrix}

\item{interactive}{logical. If \code{TRUE} prints an interactive Plotly graphic.}
}
\description{
\code{pc_pairwise_plot}
}

