% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pc_scores}
\alias{pc_scores}
\title{Reduced-rank Principal Component Scores}
\usage{
pc_scores(x, rank = "full")
}
\arguments{
\item{x}{a data frame or matrix of predictor variables.}
}
\description{
\code{pc_scores} returns data frame of principle component scores from reduced-rank PCA
}

