% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pc_scores}
\alias{pc_scores}
\title{Reduced-rank Principal Component Scores}
\usage{
pc_scores(x, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\value{
data frame with \code{rank} number of columns, each of which represent the principal component scores of the observations.
}
\description{
\code{pc_scores} returns data frame of principle component scores from reduced-rank PCA
}
\examples{
data(pendigits)
digits_features <- pendigits[, -35:-36]
pc_scores(digits_features, rank = 3)

}

