% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Fit Reduced-Rank Principal Component Model}
\usage{
pca(x, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\description{
\code{pca} fits reduced-rank principal component analysis model.
}
\examples{
data(pendigits)
digits_features <- pendigits[, -35:-36]
pca(digits_features, rank = 3)

}
\references{
Izenman, A. J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

