% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Reduced-Rank PCA}
\usage{
pca(x, rank = "full")
}
\arguments{
\item{x}{data frame of input variables}

\item{rank}{of coefficient matrix. Default \code{="full"}.}
}
\description{
\code{pca} carries out reduced-rank principled component analysis on a
matrix of input data.
}

