% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_error}
\alias{pca_error}
\title{Reduced-Rank PCA Error}
\usage{
pca_error(pca_object, x_new)
}
\arguments{
\item{pca_object}{`list` object obtained from `pca()`}

\item{x_new}{data frame or matrix of new observations to predict.}
}
\description{
\code{pca_error}
}

