% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_predict}
\alias{pca_predict}
\title{Predict via Reduced-Rank Principle Components Analysis}
\usage{
pca_predict(rrpca_object, x_new)
}
\arguments{
\item{rrpca_object}{`list` object obtained from `rrpca()`}

\item{x_new}{data frame or matrix of new observations to predict.}
}
\description{
\code{rrpca_predict}
}

