% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_predict}
\alias{pca_predict}
\title{Predict via Reduced-Rank Principal Component Analysis}
\usage{
pca_predict(pca_object, x_new)
}
\arguments{
\item{pca_object}{`list` object obtained from `pca()`}

\item{x_new}{data frame or matrix of new observations to predict.}
}
\description{
\code{pca_predict}
}

