% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{PCA Rank Trace}
\usage{
pca_rank_trace(x, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
Rank trace to determine number of principle components to use in reduced-rank PCA.
}
\examples{
x <- data.frame(x1 = runif(100, 0, 20), x2 = rnorm(100, 30, 4), x3 = rnorm(100, 10, 5), x4 = rnorm(100, 50, 15), runif(100, 0, 40))
pca_rank_trace(x, k = 0.001)

}

