% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_trace.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{Rank Trace Plot}
\usage{
pca_rank_trace(x)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_mat}{weight matrix


PCA Rank Trace

Rank trace coordinates to determine number of principle components to use in reduced-rank PCA.}

\item{x}{data frame or matrix of input variables}
}
\description{
Plot of rank trace to determine suitable rank of coefficient matrix.
}

