% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{PCA Rank Trace Plot}
\usage{
pca_rank_trace(x, type = "cov", k = 0, plot = TRUE, interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of predictor variables}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}

\item{plot}{if FALSE, returns data frame of rank trace coordinates.}

\item{interactive}{if TRUE, creates an interactive plotly graphic.}
}
\value{
ggplot object if \code{plot = TRUE}, a data frame of rank trace coordinates if \code{plot = FALSE},
or an interactive plotly object if \code{interactive = TRUE}.
}
\description{
\code{pca_rank_trace} is a plot used to determine the effective dimensionality, i.e., \eqn{t = \mathrm{rank}\left(\mathbf{C}\right)},
of the reduced-rank regression equation, or the number of principal components to be used in the model.
}
\examples{
data(pendigits)
digits_features <- pendigits[,1:34]
pca_rank_trace(digits_features)
pca_rank_trace(digits_features, plot = FALSE)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}
\seealso{
\code{\link{rank_trace}}
}

