% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{PCA Rank Trace}
\usage{
pca_rank_trace(x)
}
\arguments{
\item{x}{data frame of input variables}
}
\description{
Rank trace to determine number of principle components to use in reduced-rank PCA.
}

