% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_trace.R, R/pca_rank_trace.R
\name{pca_rank_trace}
\alias{pca_rank_trace}
\title{PCA Rank Trace}
\usage{
pca_rank_trace(x)

pca_rank_trace(x)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{x}{export pca_rank_trace}
}
\description{
Rank trace coordinates to determine number of principle components to use in reduced-rank PCA.

\code{pca_rank_trace} is a wrapper function for \code{rank_trace} for Principle Components Analysis
}

