% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_trace.R, R/pca_rank_trace.R
\name{pca_rank_trace_plot}
\alias{pca_rank_trace_plot}
\title{PCA Rank Trace Plot}
\usage{
pca_rank_trace_plot(x)

pca_rank_trace_plot(x)
}
\arguments{
\item{x}{data frame or matrix of input variables}
}
\description{
Plot of rank trace to determine number of principle components to use in reduced-rank PCA.

\code{pca_rank_trace} is a wrapper for \code{rank_trace}
}
\details{
export pca_rank_trace
}

