% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_trace.R, R/pca.R
\name{pca_rank_trace_plot}
\alias{pca_rank_trace_plot}
\title{PCA Rank Trace Plot}
\usage{
pca_rank_trace_plot(x, interactive = FALSE)

pca_rank_trace_plot(x, interactive = FALSE)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{x}{data frame or matrix of input variables}
}
\description{
Plot of rank trace to determine number of principle components to use in reduced-rank PCA.

Plot of rank trace to determine number of principle components to use in reduced-rank PCA.
}

