% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_rank_trace_plot}
\alias{pca_rank_trace_plot}
\title{PCA Rank Trace Plot}
\usage{
pca_rank_trace_plot(x, interactive = FALSE)
}
\arguments{
\item{x}{data frame of input variables}

\item{interactive}{logical. If \code{TRUE} prints an interactive Plotly graphic.}
}
\description{
Plot of rank trace to determine number of principle components to use in reduced-rank PCA.
}

