% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_rank_trace.R
\name{pca_rank_trace_plot}
\alias{pca_rank_trace_plot}
\title{Rank Trace Plot for PCA}
\usage{
pca_rank_trace_plot(x)
}
\description{
\code{pca_rank_trace} is a wrapper for \code{rank_trace}
}
\details{
export pca_rank_trace
}

