% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_plots.R
\name{rank_trace}
\alias{rank_trace}
\title{Rank Trace}
\usage{
rank_trace(x, y, gamma_matrix, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small constant to add to the ridge of covariance matrices to avoid singularity.}
}
\value{
data frame containing rank trace.
}
\description{
Rank trace to determine number of principle components to use in reduced-rank PCA.
}
\examples{
data(tobacco)
tobacco_x <- tobacco[,4:9]
tobacco_y <- tobacco[,1:3]
gamma <- diag(1, dim(tobacco_y)[2])
rank_trace(tobacco_x, tobacco_y, gamma)

}

