% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_plots.R
\name{rank_trace_plot}
\alias{rank_trace_plot}
\title{Rank Trace Plot}
\usage{
rank_trace_plot(x, y, gamma_mat)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}
}
\description{
Plot of rank trace to determine suitable rank of coefficient matrix.
}

