% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr_plots.R
\name{rank_trace_plot}
\alias{rank_trace_plot}
\title{Rank Trace Plot}
\usage{
rank_trace_plot(x, y, gamma_matrix, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
Plot of rank trace to determine suitable rank of coefficient matrix.
}

