% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{rrlda}
\alias{rrlda}
\title{Fit Reduced-Rank LDA Model}
\usage{
rrlda(x, y, k = 0, ld_num = 2)
}
\arguments{
\item{x}{data frame of input variables}

\item{y}{data frame of response variables}

\item{k}{small number to add to the ridge.}
}
\description{
\code{lda} fits a reduced-rank linear discriminant analysis model.
}
\references{
Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.
}

