% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{rrlda}
\alias{rrlda}
\title{Fit Reduced-Rank LDA Model}
\usage{
rrlda(x, y, rank, type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} is standard multivariate regression.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{lda} fits a reduced-rank linear discriminant analysis model.
}
\references{
Izenman, A. J. (2008) Modern Multivariate Statistical Techniques. Springer.
}

