% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr}
\alias{rrr}
\title{Reduced-Rank Regression}
\usage{
rrr(x, y, gamma_matrix, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame of input variables}

\item{y}{data frame of response variables}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} is standard multivariate regression.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{gamma_mat}{weight matrix}
}
\description{
\code{rrr} fits a reduced-rank regression model.
}

