% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr}
\alias{rrr}
\title{Fit Reduced-Rank Regression Model}
\usage{
rrr(x, y, gamma_matrix, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{rrr} fits a reduced-rank regression model.
}
\examples{
data(tobacco)
tobacco_x <- tobacco[,1:3]
tobacco_y <- tobacco[,4:9]
gamma <- diag(1, dim(tobacco_y)[2])
rrr(tobacco_x, tobacco_y, gamma, rank = 1)

}
\references{
Izenman, A.J. (2008) \emph{Modern Multivariate Statistical Techniques}. Springer.
}

