% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr_residuals}
\alias{rrr_residuals}
\title{RRR Residuals}
\usage{
rrr_residuals(x, y, gamma_matrix, rank = "full", type = "cov", k = 0)
}
\arguments{
\item{x}{data frame or matrix of input variables}

\item{y}{data frame or matrix of response variables}

\item{gamma_matrix}{weight matrix}

\item{rank}{of the coefficient matrix to estimate. Default \code{rank = full} prodcues the standard multivariate regression technique.}

\item{type}{the format of the covariance matrix. \code{type = "cov"} runs the regression using the mean-centered covariance matrix. \code{type = "cor"} runs the regression using the mean-centered, standard-deviation-scaled correlation matrix.}

\item{k}{small number to add to the ridge.}
}
\description{
\code{rrr_residuals}
}

